# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Furniture AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/search/examples": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Examples",
        "description": "Sentences known to work, for empty search and room-planner screens.",
        "operationId": "get_examples_v1_search_examples_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ar",
                "en"
              ],
              "type": "string",
              "default": "en",
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamplesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamplesResponse": {
        "properties": {
          "search": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Search"
          },
          "search_refinements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Search Refinements"
          },
          "rooms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rooms"
          },
          "room_refinements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Room Refinements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "search",
          "search_refinements",
          "rooms",
          "room_refinements"
        ],
        "title": "ExamplesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}